+++ server/app/controllers/AnyplaceAccounts.scala
@@ -52,43 +52,6 @@ import utils.{AnyResponseHelper, JsonUtils, LPLogger}
 object AnyplaceAccounts extends Controller {
 
   /**
-   * Retrieve all the accounts.
-   *
-   * @return
-   */
-  def fetchAllAccounts() = Action {
-    implicit request =>
-
-      def inner(request: Request[AnyContent]): Result = {
-        // TODO: Only for admin users if not admin return only for admins
-        val anyReq: OAuth2Request = new OAuth2Request(request)
-        if (!anyReq.assertJsonBody()) {
-          return AnyResponseHelper.bad_request(
-            AnyResponseHelper.CANNOT_PARSE_BODY_AS_JSON)
-        }
-
-        LPLogger.info("AnyplaceAccounts::fetchAllAccounts(): ")
-        try {
-          return AnyResponseHelper.bad_request("User is not admin.")
-          val users: List[JsValue] = ProxyDataSource.getIDatasource().getAllAccounts()
-          val res: JsValue = Json.obj(
-            "users_num" -> users.length,
-            SCHEMA.cUsers -> Json.arr(users)
-          )
-          AnyResponseHelper.ok(res, "Successfully retrieved all accounts!")
-        } catch {
-          case e: DatasourceException =>
-            return AnyResponseHelper.internal_server_error(
-              "500: " + e.getMessage)
-
-        }
-      }
-
-      inner(request)
-  }
-
-
-  /**
    * Fetches the account with the AUID passed in.
    * The account document is returned in the Json response.
    *
